+++ src/command_modules/azure-cli-role/azure/cli/command_modules/role/_help.py
@@ -7,33 +7,85 @@ from azure.cli.core.help_files import helps  # pylint: disable=unused-import
 
 # pylint: disable=line-too-long
 
+
 helps['ad sp create-for-rbac'] = """
-            examples:
-                - name: Create with a default role assignment.
-                  text: az ad sp create-for-rbac
-                - name: Create using a custom name, and with a default assiggment.
-                  text: az ad sp create-for-rbac -n "http://MyApp"
-                - name: Create without a default assignment.
-                  text: az ad sp create-for-rbac --skip-assignment
-                - name: Create with customized assignments
-                  text: az ad sp create-for-rbac -n "http://MyApp" --role contributor --scopes /subscriptions/11111111-2222-3333-4444-555555555555/resourceGroups/MyResourceGroup /subscriptions/11111111-2222-3333-4444-666666666666/resourceGroups/MyAnotherResourceGroup
-                - name: Create using self-signed certificte
-                  text: az ad sp create-for-rbac --create-cert
-                - name: Login with a service principal.
-                  text: az login --service-principal -u <name> -p <password> --tenant <tenant>
-                - name: Login with self-signed certificate
-                  text: az login --service-principal -u <name> -p <certificate file path> --tenant <tenant>
-                - name: Reset credentials on expiration.
-                  text: az ad sp reset-credentials --name <name>
-                - name: Create extra role assignments in future.
-                  text: az role assignment create --assignee <name> --role Contributor
-                - name: Revoke the service principal when done with it.
-                  text: az ad app delete --id <name>
-                - name: Create using certificate from Key Vault
-                  text: >
-                    az keyvault certificate download --vault-name vault -n cert-name -f cert.pem \n
-                    az ad sp create-for-rbac --cert @cert.pem
-            """
+    type: command
+    short-summary: Create a service principal and configure its access to Azure resources.
+    parameters:
+        - name: --name -n
+          short-summary: Display name or an app ID URI. Command will generate one if missing.
+        - name: --password -p
+          short-summary: The password used to login. If missing, command will generate one.
+        - name: --cert
+          short-summary: Certificate to use for credentials in lieu of password.
+          long-summary: When using --key-vault, indicates the name of the cert to use or create.
+            Otherwise, supply a PEM or DER formatted public certificate string. Use `@<file path>` to
+            load from a file. Do not include private key info.
+        - name: --create-cert
+          short-summary: Create a self-signed certificate to use for the credential.
+          long-summary: Use with --key-vault to create the certificate in Key Vault. Otherwise, a
+            certificate will be created locally.
+        - name: --keyvault
+          short-summary: Name or ID of a KeyVault to use for creating or retrieving certificates.
+        - name: --years
+          short-summary: >
+            Number of years for which the credentials will be valid. Default: 1 year
+        - name: --scopes
+          short-summary: Space-separated list of scopes the service principal's role assignment applies to.
+            Defaults to the root of the current subscription.
+        - name: --role
+          short-summary: Role the service principal has in regard to resources.
+    examples:
+        - name: Create with a default role assignment.
+          text: az ad sp create-for-rbac
+        - name: Create using a custom name, and with a default assiggment.
+          text: az ad sp create-for-rbac -n "http://MyApp"
+        - name: Create without a default assignment.
+          text: az ad sp create-for-rbac --skip-assignment
+        - name: Create with customized assignments
+          text: az ad sp create-for-rbac -n "http://MyApp" --role contributor --scopes /subscriptions/11111111-2222-3333-4444-555555555555/resourceGroups/MyResourceGroup /subscriptions/11111111-2222-3333-4444-666666666666/resourceGroups/MyAnotherResourceGroup
+        - name: Create using self-signed certificte
+          text: az ad sp create-for-rbac --create-cert
+        - name: Create self-signed certificate within KeyVault
+          text: az ad sp create-for-rbac --key-vault <vault name> --cert <name> --create-cert
+        - name: Create using existing certificate in KeyVault
+          text: az ad sp create-for-rbac --key-vault <vault name> --cert <name>
+        - name: Login with a service principal.
+          text: az login --service-principal -u <name> -p <password> --tenant <tenant>
+        - name: Login with self-signed certificate
+          text: az login --service-principal -u <name> -p <certificate file path> --tenant <tenant>
+        - name: Reset credentials on expiration.
+          text: az ad sp reset-credentials --name <name>
+        - name: Create extra role assignments in future.
+          text: az role assignment create --assignee <name> --role Contributor
+        - name: Revoke the service principal when done with it.
+          text: az ad app delete --id <name>
+    """
+
+helps['ad sp reset-credentials'] = """
+    type: command
+    short-summary: Reset service principal credentials.
+    long-summary: Use upon expiration of the existing credentials or in the even that you forget them.
+    parameters:
+        - name: --name -n
+          short-summary: Display name or an app ID URI.
+        - name: --password -p
+          short-summary: The password used to login. If missing, command will generate one.
+        - name: --cert
+          short-summary: Certificate to use for credentials in lieu of password.
+          long-summary: When using --key-vault, indicates the name of the cert to use or create.
+            Otherwise, supply a PEM or DER formatted public certificate string. Use `@<file path>` to
+            load from a file. Do not include private key info.
+        - name: --create-cert
+          short-summary: Create a self-signed certificate to use for the credential.
+          long-summary: Use with --key-vault to create the certificate in Key Vault. Otherwise, a
+            certificate will be created locally.
+        - name: --keyvault
+          short-summary: Name or ID of a KeyVault to use for creating or retrieving certificates.
+        - name: --years
+          short-summary: >
+            Number of years for which the credentials will be valid. Default: 1 year
+"""
 
 helps['role'] = """
     type: group