+++ src/command_modules/azure-cli-role/HISTORY.rst
@@ -3,6 +3,15 @@
 Release History
 ===============
 
+unreleased
+++++++++++++++++++
+* Add Key Vault integration to the create-for-rbac and reset-credentials commands.
+
+
+2.0.4 (2017-05-09)
+++++++++++++++++++
+* Minor fixes.
+
 2.0.3 (2017-04-28)
 ++++++++++++++++++
 * create-for-rbac: ensure SP's end date will not exceed certificate's expiration date (#2989)